refactor: move finalizers and requeueu to independent files and cleanup errors
