+++ controllers/errors.go
@@ -3,9 +3,7 @@ package controllers
 import (
 	"errors"
 	"fmt"
-	"time"
 
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
 
@@ -15,16 +13,8 @@ var (
 
 // TypeConvertError returns an error indicating a generic controllers.Request interface
 // could not be converted to its underlying type.
-func TypeConvertError(requeue time.Duration, t interface{}) (ctrl.Result, error) {
-	return RequeueAfter(requeue), fmt.Errorf(
-		"unable to convert controllers.Request interface to underlying request type [%T]",
-		t,
-	)
-}
-
-// FinalizerError returns an error when attempting to act upon a finalizer.
-func FinalizerError(requeue time.Duration, err error) (ctrl.Result, error) {
-	return RequeueAfter(requeue), err
+func TypeConvertError(t interface{}) error {
+	return fmt.Errorf("unable to convert controllers.Request interface to underlying request type [%T]", t)
 }
 
 // AddFinalizerError returns an error when registering a finalizer.
