+++ controllers/gitlabidentityprovider/phases.go
@@ -225,7 +225,7 @@ func (r *Controller) Complete(request *GitLabIdentityProviderRequest) (ctrl.Resu
 // CompleteDestroy will perform all actions required to successfully complete a delete reconciliation request.
 func (r *Controller) CompleteDestroy(request *GitLabIdentityProviderRequest) (ctrl.Result, error) {
 	if err := controllers.RemoveFinalizer(request.Context, r, request.Original); err != nil {
-		return controllers.FinalizerError(defaultGitLabIdentityProviderRequeue, controllers.RemoveFinalizerError(err))
+		return controllers.RequeueOnError(request, controllers.RemoveFinalizerError(err))
 	}
 
 	request.Log.Info("completed gitlab identity provider deletion", controllers.LogValues(request)...)
