+++ controllers/rosacluster/phases.go
@@ -321,7 +321,7 @@ func (r *Controller) Complete(request *ROSAClusterRequest) (ctrl.Result, error) 
 // CompleteDestroy will perform all actions required to successfully complete a delete reconciliation request.
 func (r *Controller) CompleteDestroy(request *ROSAClusterRequest) (ctrl.Result, error) {
 	if err := controllers.RemoveFinalizer(request.Context, r, request.Original); err != nil {
-		return controllers.FinalizerError(defaultClusterRequeue, controllers.RemoveFinalizerError(err))
+		return controllers.RequeueOnError(request, controllers.RemoveFinalizerError(err))
 	}
 
 	request.Log.Info("completed rosa cluster deletion", controllers.LogValues(request)...)
